requirements.txt
torch>=2.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
numpy>=1.21.0
tqdm>=4.65.0